back/app/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
pydantic[email]
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart